+++ src/FuryLionConfrontation/Assets/Code/StaticData/Balance/IBalanceEntry.cs
@@ -6,7 +6,7 @@ namespace Confrontation
 
 		public interface IGoldenMine
 		{
-			int ProduceCollDownDuration { get; }
+			float ProduceCollDownDuration { get; }
 
 			int ProduceAmount { get; }
 		}
